+++ src/IGDBWrappers/AddIGDBScores.cs
@@ -23,7 +23,7 @@ namespace GameListDB.IGDBWrappers
                 System.Console.WriteLine("Getting Score of " + game.Name);
                 gameList = await igdb.RequestQuery("id,rating", gamelistdb.GetIgdbId(game));
                 if (gameList.Length == 0 || !gameList.FirstOrDefault().Rating.HasValue) { System.Console.WriteLine("It doesn't have score"); continue; }
-                game.Score = (long)gameList.FirstOrDefault().Rating;
+                game.Score = (int)gameList.FirstOrDefault().Rating;
                 gamelistdb.Backlogs.Update(game);
 
             }
